Let port/stub/dsound.h own the DirectSound names lib/sound.cpp needs, and make HRESULT 32 bits so FAILED() works.
docs/porting/audio-seams.md wrote down the trigger for this: grow the stub
when lib/sound.cpp or lib/wave.cpp is allowlisted. Only the second half had
happened, and #86 met it by giving lib/wave.cpp a private copy of the names
instead of growing the header, so the trigger stayed half-armed and
lib/sound.cpp stayed off the allowlist behind twelve missing declarations.

Arming it fully means the private copy has to go. Its #ifndef DSBUFFERDESC
guard cannot fire -- DSBUFFERDESC is a typedef, not a macro -- so declaring
DSBUFFERDESC in the stub turns lib/wave.cpp, already allowlisted, into
"typedef redefinition with different types". The block is not upstream:
2662b8a:lib/wave.cpp has no such lines and git log -S names ddf1721 (#86) as
the commit that added them. Deleting it is port cleanup, not a game-code
rewrite, and lib/wave.cpp's statements are untouched.

HRESULT is in this commit because DirectSoundCreate8 is what exposed it.
Returning DS_OK from a stub that produces no device sends InitDirectSound past
the "svs.pDS = NULL; return TRUE" branch the game wrote for a machine with no
sound, into SetCooperativeLevel through a null svs.pDS. E_FAIL should have
fixed that and did not: HRESULT was long, so on LP64 every 0x8....... constant
stayed positive and FAILED() answered false for E_FAIL, E_NOTIMPL, every
DSERR_* and every D3DERR_*. SUCCEEDED(E_FAIL) was true. Nothing failed to
compile, which is why it survived -- port/ffp_fvf_test.cpp had already hit it
and worked around it in the test rather than in the type, and the soft-fail
dshow.h documents for InitDirectShow was broken the whole time. That test now
asserts through FAILED(), so it fails if the width regresses.

#124 also carried lib/sound.cpp in its "needs a real backend" row. It did
not: nothing here implements DirectSound, and the stub bodies stay the
return S_OK shape that #7 will replace.

Closes #152
Parent: #124

+++ port/stub/windows.h
@@ -272,7 +272,12 @@ typedef LRESULT(CALLBACK *WNDPROC)(HWND, UINT, WPARAM, LPARAM);
 #define E_NOTIMPL ((HRESULT)0x80004001L)
 #endif
 
-typedef long HRESULT;
+// int, not long: Windows HRESULT is 32 bits, and on LP64 a 64-bit HRESULT
+// makes every 0x8....... constant positive, so FAILED() answers false for
+// E_FAIL, E_NOTIMPL and every DSERR_*. Soft-fail paths written against
+// FAILED() -- InitDirectSound here, InitDirectShow in dshow.h -- would run on
+// as if the call had succeeded.
+typedef int HRESULT;
 #define FAILED(hr) (((HRESULT)(hr)) < 0)
 #define SUCCEEDED(hr) (((HRESULT)(hr)) >= 0)
